+++ app/views/api/v1/commission_histories/_commission_history.json.jbuilder
@@ -1,0 +1,8 @@
+# frozen_string_literal: true
+
+json.account              commission_history.validator.account
+json.epoch                commission_history.epoch
+json.commission_before    commission_history.commission_before
+json.commission_after     commission_history.commission_after
+json.network              commission_history.network
+json.timestamp            commission_history.created_at
